Remove Events v1beta specific code
