+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/resource/ResourceOperatorSupplier.java
@@ -5,15 +5,16 @@
 package io.strimzi.operator.cluster.operator.resource;
 
 import io.fabric8.kubernetes.client.KubernetesClient;
+import io.fabric8.openshift.client.OpenShiftClient;
 import io.strimzi.api.kafka.KafkaBridgeList;
 import io.strimzi.api.kafka.KafkaConnectList;
 import io.strimzi.api.kafka.KafkaConnectorList;
-import io.strimzi.api.kafka.KafkaMirrorMakerList;
-import io.strimzi.api.kafka.KafkaMirrorMaker2List;
-import io.strimzi.api.kafka.KafkaRebalanceList;
 import io.strimzi.api.kafka.KafkaList;
-import io.strimzi.api.kafka.model.KafkaBridge;
+import io.strimzi.api.kafka.KafkaMirrorMaker2List;
+import io.strimzi.api.kafka.KafkaMirrorMakerList;
+import io.strimzi.api.kafka.KafkaRebalanceList;
 import io.strimzi.api.kafka.model.Kafka;
+import io.strimzi.api.kafka.model.KafkaBridge;
 import io.strimzi.api.kafka.model.KafkaConnect;
 import io.strimzi.api.kafka.model.KafkaConnector;
 import io.strimzi.api.kafka.model.KafkaMirrorMaker;
@@ -45,8 +46,6 @@ import io.strimzi.operator.common.operator.resource.RouteOperator;
 import io.strimzi.operator.common.operator.resource.SecretOperator;
 import io.strimzi.operator.common.operator.resource.ServiceAccountOperator;
 import io.strimzi.operator.common.operator.resource.ServiceOperator;
-
-import io.fabric8.openshift.client.OpenShiftClient;
 import io.strimzi.operator.common.operator.resource.StorageClassOperator;
 import io.strimzi.operator.common.operator.resource.StrimziPodSetOperator;
 import io.vertx.core.Vertx;
@@ -100,7 +99,7 @@ public class ResourceOperatorSupplier {
                 metricsProvider,
                 pfa,
                 operationTimeoutMs,
-                KubernetesRestartEventPublisher.createPublisher(client, operatorName, pfa.hasEventsApiV1())
+                new KubernetesRestartEventPublisher(client, operatorName)
         );
     }
 
@@ -121,7 +120,8 @@ public class ResourceOperatorSupplier {
                 metricsProvider,
                 pfa,
                 operationTimeoutMs,
-                KubernetesRestartEventPublisher.createPublisher(client, "operatorName", pfa.hasEventsApiV1()));
+                new KubernetesRestartEventPublisher(client, "operatorName")
+        );
     }
 
     //Exposed for testing
